Clump board server bits together.
